refactor(server): extract startup recovery to src/core/server/startup-recovery.ts
Second safe slice of the server.ts split. Moves the post-listen recovery IIFE — startup
sweep (markPriorBootRunsLost) → completion backfill → restart recovery → dispatch.start +
watchdog.start (in a finally) → post-recovery dispatchTasks → doctor.start (in a finally) →
BAKIN_SEED_USAGE — into src/core/server/startup-recovery.ts as runStartupRecovery(contentDir,
port). The listen callback now just calls `void runStartupRecovery(CONTENT_DIR, port)`.

Every ordering invariant is preserved verbatim (sweep BEFORE recovery; dispatch/watchdog
start even if recovery throws; doctor in the trailing finally); the dev usage-seed dynamic
import stays dynamic (graph-pruning). watchdog/doctor namespace imports drop out of server.ts
(now only used inside the recovery module). server.ts: 777 → 727.

Verified: typecheck + lint clean; full suite 5105/0; full 3-target binary build; and an
isolated-home boot smoke — the built binary boots and the recovery sequence runs through the
extracted path: Dispatch / Watchdog / Doctor all started in order (sweep/recovery log lines
correctly silent on a fresh empty home with nothing to sweep or recover).
